refactor(bulk_sms): Remove 'Send Immediately' option
The "Send Immediately" checkbox has been removed from the Bulk SMS doctype to simplify the user interface and sending logic.

The `scheduled_datetime` field is now always visible and is the sole method for determining the send time. To send a message immediately, users can set the scheduled time to the current date and time. This change streamlines the process and reduces code complexity.